refactor: Standardize component file naming to PascalCase
- Rename component files from kebab-case to PascalCase convention
- Update import/export paths in index.ts files to match new file names
- Maintain consistent naming across components, icons, navigation, theme, and utility modules
