+++ src/components/icons/AdaptiveIcon.tsx
