+++ src/components/index.ts
@@ -1,14 +1,14 @@
 // Navigation Components
-export { default as Navbar } from './nav/navbar';
-export { default as MobileNav } from './nav/mobile-nav';
-export { default as Footer } from './nav/footer';
-export { ThemeToggle } from './theme/theme-toggle';
+export { default as Navbar } from './nav/NavBar';
+export { default as MobileNav } from './nav/MobileNav';
+export { default as Footer } from './nav/Footer';
+export { ThemeToggle } from './theme/ThemeToggle';
 
 // UI Components
 export { Button } from './ui/button';
-export { SocialLinks } from './social-links';
-export { WordRotate } from './magicui/word-rotate';
+export { SocialLinks } from './SocialLinks';
+export { WordRotate } from './magicui/WordRotate';
 
 // Utility Components
-export { DynamicThemeColor } from './theme/dynamic-theme-color';
-export { DynamicViewport } from './dynamic-viewport'; 
+export { DynamicThemeColor } from './theme/DynamicThemeColor';
+export { DynamicViewport } from './DynamicViewport';
