+++ src/components/magicui/WordRotate.tsx
